Changed 2FA verification input for the settings modal
