+++ client/src/components/tools/Interface.ts
@@ -8,7 +8,6 @@ export interface data {
     // picture: event.target.picture.value,
 }
 
-
 export interface statObj {
     valid: boolean;
     touched: boolean;
@@ -18,3 +17,7 @@ export interface validFunc {
     (e:React.ChangeEvent<HTMLInputElement>) : void;
   }
   
+
+export interface handleKeyUp {
+    handleKeyUp(field1: string, len: number, field2: string): () => void;
+  }
