+++ client/src/styles/globals.css
@@ -4,16 +4,6 @@
 @tailwind components;
 @tailwind utilities;
 
-@layer base {
-        input[type='number']::-webkit-outer-spin-button,
-        input[type='number']::-webkit-inner-spin-button,
-        input[type='number'] {
-          -webkit-appearance: none;
-          margin: 0;
-          -moz-appearance: textfield !important;
-        }
-}
-
 html {
   min-width: 300px;
   min-height: 750px;
